requirements.txt
beautifulsoup4==4.4.1
Django==1.9.2
django-appconf==1.0.1
django-wysiwyg-redactor==0.4.9
Pillow==3.1.1
psycopg2==2.6.1
six==1.10.0
wheel==0.26.0
